2022.02.06 Bong

+++ src/main/java/com/mumulcom/mumulcom/src/user/service/UserService.java
@@ -124,6 +124,9 @@ public class UserService {
         if (userOptional.isEmpty()) {
             throw new BaseException(BaseResponseStatus.RESPONSE_ERROR);
         }
+        if (existsByNickname(patchReq.getNickname())) {
+            throw new BaseException(BaseResponseStatus.POST_USERS_EXISTS_NICKNAME);
+        }
         User user = userOptional.get();
         user.updateUserInfo(
                 patchReq.getNickname(),
@@ -131,7 +134,7 @@ public class UserService {
                 patchReq.getProfileImgUrl()
         );
         //관심 코딩 분야에서 변경된 사항이 있으면
-        if (isMyCategoriesChanged(getMyCategoriesByUserIdx(user.getUserIdx()), patchReq.getMyCategories())) {
+        if (patchReq.getMyCategories() != null && isMyCategoriesChanged(getMyCategoriesByUserIdx(user.getUserIdx()), patchReq.getMyCategories())) {
             //기존 레코드 삭제
             myCategoryRepository.deleteAll(myCategoryRepository.findByUserIdx(user.getUserIdx()));
             //새로 저장
